+++ src/main/kotlin/com/demo/book/movie/request/ShowRequest.kt
@@ -1,8 +1,10 @@
 package com.demo.book.movie.request
 
+import java.math.BigDecimal
+
 data class ShowRequest(
     val startTime: Long,
-    val movieId : Int,
-    val price : Int
+    val movieId: Int,
+    val price: Double
 )
